delete accountの実装

+++ api/conf/app_user/views.py
@@ -8,7 +8,7 @@ from rest_framework.views import APIView
 from django.utils.decorators import method_decorator
 from django.db import transaction, DatabaseError
 from django.views import View
-from rest_framework_simplejwt.tokens import RefreshToken
+from rest_framework_simplejwt.tokens import AccessToken, RefreshToken
 from django.http import JsonResponse
 from .models import User
 from django.contrib.auth.hashers import make_password
@@ -16,7 +16,6 @@ import json
 from django.http import JsonResponse
 from rest_framework import status
 from rest_framework.views import APIView
-
 
 
 class UserLoginView(APIView):
@@ -72,15 +71,13 @@ class UserView(APIView):
                 }, status=status.HTTP_400_BAD_REQUEST)
 
             # ユーザーを作成
-            user = User.objects.create(
+            user = User.objects.create_user(
                 login_name=login_name,
-                password=make_password(password=password, salt='ft_transcendence'),
+                password=password,
                 display_name=display_name,
                 secret_question=secret_question,
-                secret_answer_hash=make_password(secret_answer, salt='ft_transcendence'),
+                secret_answer=secret_answer
             )
-            user.user_id = user.id
-            user.save()
 
 
             return JsonResponse({
@@ -97,18 +94,39 @@ class UserView(APIView):
 
     def delete(self, request, *args, **kwargs):
         try:
-            login_name = json.loads(request.body).get('login_name')
+            data = json.loads(request.body)
+            login_name = data.get('login_name')
+            auth = request.headers.get('Authorization')
+
+            if auth:
+                access_token = auth.split(' ')[1]
+            if not access_token:
+                return JsonResponse({
+                    'message': 'Login before you delete your account'
+                }, status=status.HTTP_401_UNAUTHORIZED)
+
+
+            access_id = AccessToken(access_token).get('user_id')
             user = User.objects.get(login_name=login_name)
+
+
+            if (user.id != access_id):
+                return JsonResponse({
+                    'message': 'You can only delete your own account'
+                }, status=status.HTTP_400_BAD_REQUEST)
+
+
             user.delete()
+
             return JsonResponse({
                 'message': 'User deleted.'
             }, status=status.HTTP_200_OK)
+
         except User.DoesNotExist:
             return JsonResponse({
                 'error': 'User not found.'
             }, status=status.HTTP_404_NOT_FOUND)
         except Exception as e:
-
             return JsonResponse({
                 'error': str(e)
